Add new transportation object
So much transformation needs to be done to the transportation
information that I thought it would be best to create a new object for
it. This object will be used by both single activities and activity
groups to build raw data structures.

JavaScript and HTML construction will happen at the shortcode level in
the PonoRezTemplate class.

Addresses #7